% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_path}
\alias{add_path}
\title{Overlay paths of interpolation on the sphere}
\usage{
add_path(proj_space, path)
}
\arguments{
\item{proj_space}{n number of points on the surface of sphere}

\item{path}{interpolated path}
}
\description{
Overlay paths of interpolation on the sphere
}
