% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{calculate_tau}
\alias{calculate_tau}
\title{Compute angle between two d-dimensional frames in p-space}
\usage{
calculate_tau(F1, F2)
}
\arguments{
\item{F1}{pxd frame}

\item{F2}{pxd frame}
}
\value{
tau angle in radians
}
\description{
Compute angle between two d-dimensional frames in p-space
}
\examples{
calculate_tau(Wz, Wa)
}
