% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{construct_rotation_matrix}
\alias{construct_rotation_matrix}
\title{Construct rotation matrix in given angle}
\usage{
construct_rotation_matrix(theta)
}
\arguments{
\item{theta}{angle  of rotation}
}
\value{
rotation matrix
}
\description{
Construct rotation matrix in given angle
}
\examples{

theta <- 1
construct_rotation_matrix(theta)
}
