% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{construct_Wa}
\alias{construct_Wa}
\title{Construct starting basis from pre-projection}
\usage{
construct_Wa(Fa, B)
}
\arguments{
\item{Fa}{starting pxd frame}

\item{B}{pre-projection px2d matrix}
}
\value{
Wa starting 2dxd frame on preprojection space (first dxd entry of this matrix is identity matrix by construction)
}
\description{
Construct starting basis from pre-projection
}
\examples{
construct_Wa(base1, b) 
}
