% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{construct_Wz}
\alias{construct_Wz}
\title{Construct target basis from pre-projection}
\usage{
construct_Wz(Fz, B)
}
\arguments{
\item{Fz}{target pxd frame}

\item{B}{pre-projection px2d matrix}
}
\value{
Wz target 2dxd frame on preprojection space
}
\description{
Construct target basis from pre-projection
}
\examples{
construct_Wz(base2, b)
}
