% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_space_view}
\alias{generate_space_view}
\title{Generate the background sphere or torus}
\usage{
generate_space_view(n = 1000, p = 3, d = 1)
}
\arguments{
\item{n}{number of points on the sphere}

\item{p}{dimension of data}

\item{d}{dimension of projection}
}
\value{
n number of points on the surface of sphere
}
\description{
Generate the background sphere or torus
}
\examples{
p <- 4
sp <- generate_space_view(p=p)
}
