% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{givens_full_path}
\alias{givens_full_path}
\title{Construct full interpolated frames}
\usage{
givens_full_path(B, Wa, tau, nsteps)
}
\arguments{
\item{B}{pre-projection px2d matrix}

\item{Wa}{starting basis in pre-projected space}

\item{tau}{angle between starting and target basis}

\item{nsteps}{number of steps of interpolation}
}
\value{
return array with nsteps matrix. Each matrix is interpolated frame in between starting and target frames.
}
\description{
Construct full interpolated frames
}
\examples{
p <- 4
base1 <- tourr::orthonormalise(tourr::basis_random(p, d=1))
base2 <- tourr::orthonormalise(tourr::basis_random(p, d=1))
b <- preprojection(base1, base2)
Wa <- construct_preframe(base1, b) 
Wz <- construct_preframe(base2, b) 
tau <- calculate_tau(Wz, Wa)
givens_full_path(b, Wa, tau, nsteps=10)
}
