% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{givens_info}
\alias{givens_info}
\title{Calculate information needed for Givens interpolation}
\usage{
givens_info(Fa, Fz)
}
\arguments{
\item{Fa}{starting frame, will be orthonormalized if necessary}

\item{Fz}{target frame, will be orthonormalized if necessary}
}
\description{
The methodology is outlined in
\url{http://www-stat.wharton.upenn.edu/~buja/PAPERS/paper-dyn-proj-algs.pdf}
}
\keyword{internal}
