% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{givens_path}
\alias{givens_path}
\title{construct tour path}
\usage{
givens_path(Wa, tau, stepfraction)
}
\arguments{
\item{Wa}{starting basis in pre-projected space}

\item{tau}{angle between starting and target basis}

\item{stepfraction}{stepfraction in each interpolation}
}
\value{
A givens path by stepfraction in pre-projected space
}
\description{
construct tour path
}
\examples{
p <- 4
base1 <- tourr::basis_random(p, d=1)
base2 <- tourr::basis_random(p, d=1)
b <- preprojection(base1, base2)
Wa <- construct_preframe(base1, b) 
Wz <- construct_preframe(base2, b) 
tau <- calculate_tau(Wz, Wa)
givens_path(Wa, tau, stepfraction=0.1)
}
