% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{givens_step_fraction}
\alias{givens_step_fraction}
\title{Step along a Givens interpolated path by fraction of path length.}
\usage{
givens_step_fraction(interp, fraction)
}
\arguments{
\item{interp}{interpolated path}

\item{fraction}{fraction of distance between start and end frames}
}
\description{
Step along a Givens interpolated path by fraction of path length.
}
\keyword{internal}
