% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens_path.R
\name{grand_tour_givens}
\alias{grand_tour_givens}
\title{Create a grand tour with Givens interpolation}
\usage{
grand_tour_givens(d = 2, ...)
}
\arguments{
\item{d}{dimension of projection}

\item{...}{additional parameters to pass through}
}
\value{
creates grand tour
}
\description{
Create a grand tour with Givens interpolation
}
\examples{
data(sine_curve)
tourr::animate(sine_curve, grand_tour_givens(), tourr::display_xy())
}
