% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens_path.R
\name{guided_tour_givens}
\alias{guided_tour_givens}
\title{Create a guided tour with Givens interpolation}
\usage{
guided_tour_givens(
  index_f,
  d = 2,
  alpha = 0.5,
  cooling = 0.99,
  max.tries = 25,
  max.i = Inf,
  search_f = search_geodesic,
  n_sample = 100,
  ...
)
}
\arguments{
\item{index_f}{the index function to optimise.}

\item{d}{target dimensionality}

\item{alpha}{the initial size of the search window, in radians}

\item{cooling}{the amount the size of the search window should be adjusted
by after each step}

\item{max.tries}{the maximum number of unsuccessful attempts to find
a better projection before giving up}

\item{max.i}{the maximum index value, stop search if a larger value is found}

\item{search_f}{the search strategy to use: \code{\link{search_geodesic}}, \code{\link{search_better}},
\code{\link{search_better_random}}, \code{\link{search_polish}}. Default is \code{\link{search_geodesic}}.}

\item{n_sample}{number of samples to generate if \code{search_f} is \code{\link{search_polish}}}

\item{...}{arguments sent to the search_f}
}
\description{
Create a guided tour with Givens interpolation
}
\examples{
data(sine_curve)
tourr::animate(sine_curve, guided_tour_givens(tourr::splines2d()), tourr::display_xy())
}
\seealso{
\code{\link{cmass}}, \code{\link{holes}} and \code{\link{lda_pp}}
for examples of index functions.  The function should take a numeric
matrix and return a single number, preferably between 0 and 1.
\code{\link{search_geodesic}}, \code{\link{search_better}},
\code{\link{search_better_random}} for different search strategies
}
