% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens_path.R
\name{new_givens_path}
\alias{new_givens_path}
\title{Path needed for tour with Givens interpolation}
\usage{
new_givens_path(name, generator, frozen = NULL, ...)
}
\arguments{
\item{name}{name to give tour path}

\item{generator}{basis generator function}

\item{frozen}{matrix giving frozen variables, as described in
\code{freeze}}
}
\value{
creates path for Givens interpolation
}
\description{
Path needed for tour with Givens interpolation
}
\keyword{internal}
