% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{row_rot}
\alias{row_rot}
\title{Takes i and k th row of a matrix and rotate matrix by theta angle (requires matrix a to we 2*q matrix)}
\usage{
row_rot(a, i, k, theta)
}
\arguments{
\item{a}{matrix}

\item{i}{row}

\item{k}{row that we want to zero the element}

\item{theta}{angle between them}
}
\value{
rotated matrix a
}
\description{
Takes i and k th row of a matrix and rotate matrix by theta angle (requires matrix a to we 2*q matrix)
}
\examples{

p <- 4
base1 <- tourr::basis_random(p, d=1)
base2 <- tourr::basis_random(p, d=1)
b <- preprojection(base1, base2)
Wa <- construct_preframe(base1, b) 
Wz <- construct_preframe(base2, b) 
x1 <- as.matrix(c(Wa[1, 1], Wa[2, 1]))
y1 <- as.matrix(c(Wz[1, 1], Wz[2, 1]))
theta1 <- angle2(x1, y1)
row_rot(Wz, 1, 2, theta1)
}
