% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sine_curve measurements}
\alias{sine_curve measurements}
\alias{sine_curve}
\title{Simulated 6D data with a sine curve}
\format{
A 500x6 data frame
}
\description{
The data has 6 columns, labelled V1-V6,
where the sine curve is in V5, V6. The
other columns are normal samples.
}
\examples{
library(woylier)
data(sine_curve)
plot(sine_curve$V5, sine_curve$V6)
}
\keyword{datasets}
