% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{angle2}
\alias{angle2}
\title{Compute the angle between 2d vectors 360 degrees}
\usage{
angle2(x, y)
}
\arguments{
\item{x}{vector with length 2}

\item{y}{vector with length 2}
}
\value{
angle in radians
}
\description{
Compute the angle between 2d vectors 360 degrees
}
\keyword{internal}
