% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{calculate_angles}
\alias{calculate_angles}
\title{Calculate angles of required rotations to map Wz to Wa}
\usage{
calculate_angles(Wa, Wz)
}
\arguments{
\item{Wa}{starting preprojected frame}

\item{Wz}{target preprojected frame}
}
\value{
named list of angles
}
\description{
Calculate angles of required rotations to map Wz to Wa
}
\keyword{internal}
