% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{construct_moving_frame}
\alias{construct_moving_frame}
\title{Reconstruct interpolated frames using pre-projection}
\usage{
construct_moving_frame(Wt, B)
}
\arguments{
\item{Wt}{A givens path by stepfraction}

\item{B}{pre-projection px2d matrix}
}
\value{
A frame of on the step of interpolation
}
\description{
Reconstruct interpolated frames using pre-projection
}
\examples{
p <- 4
base1 <- tourr::basis_random(p, d=1)
base2 <- tourr::basis_random(p, d=1)
b <- preprojection(base1, base2)
Wa <- construct_preframe(base1, b) 
Wz <- construct_preframe(base2, b) 
angles <- calculate_angles(Wa, Wz)
Wt <- givens_rotation(Wa, angles, stepfraction=0.1)
construct_moving_frame(Wt, b)
}
