% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{construct_moving_frame}
\alias{construct_moving_frame}
\title{Reconstruct interpolated frames using pre-projection}
\usage{
construct_moving_frame(Wt, B)
}
\arguments{
\item{Wt}{A givens path by stepfraction}

\item{B}{pre-projection px2d matrix}
}
\value{
A frame of on the step of interpolation
}
\description{
Reconstruct interpolated frames using pre-projection
}
\keyword{internal}
