% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{construct_preframe}
\alias{construct_preframe}
\title{Construct preprojected frames}
\usage{
construct_preframe(Fr, B)
}
\arguments{
\item{Fr}{Orthonormal frame}

\item{B}{pre-projection px2d matrix}
}
\value{
Preprojected 2dxd frame on preprojection space (first dxd entry of this matrix is identity matrix by construction)
}
\description{
Construct preprojected frames
}
\keyword{internal}
