% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{givens_path}
\alias{givens_path}
\title{Interpolation format for tourr}
\usage{
givens_path(current, target, frozen = NULL, ...)
}
\arguments{
\item{current}{starting frame}

\item{target}{target frame}

\item{frozen}{indicator whether some dimensions fixed}

\item{...}{arguments sent to later functions}
}
\value{
\item{interpolate}{A function with single parameter in [0, 1] that
returns an interpolated frame between the current and future frames.
0 gives the current plane, 1 gives the new target frame in plane of
current frame.}
\item{dist}{The distance, in radians, between the current and target
frames.}
\item{Fa}{The current frame.}
\item{Fz}{The new target frame.}
\item{tau}{The principle angles between the current and target frames.}
\item{Ga}{The current plane.}
\item{Gz}{The target plane.}
}
\description{
Generates the interpolation in the form needed for
tourr, modelled on geodesic_path()
}
\keyword{internal}
