% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{givens_rotation}
\alias{givens_rotation}
\title{It implements series of Givens rotations that maps Wa to Wz}
\usage{
givens_rotation(Wa, angles, stepfraction)
}
\arguments{
\item{Wa}{starting preprojected frame}

\item{angles}{angles of required rotations to map Wz to Wa}

\item{stepfraction}{for the interpolation of rotations}
}
\value{
Givens path by stepfraction in pre-projected space
}
\description{
It implements series of Givens rotations that maps Wa to Wz
}
\keyword{internal}
