% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens_path.R
\name{planned_tour_givens}
\alias{planned_tour_givens}
\title{A planned tour path using frame-to-frame interpolation.}
\usage{
planned_tour_givens(basis_set, cycle = FALSE)
}
\arguments{
\item{basis_set}{the set of bases as a list of projection matrices
or a 3d array}

\item{cycle}{cycle through continuously (\code{TRUE}) or stop after
first pass (\code{FALSE})}
}
\value{
creates planned tour path
}
\description{
The planned tour takes you from one basis to the next in a
set order.  Once you have visited all the planned bases, you either stop
or start from the beginning once more (if \code{cycle = TRUE}).
}
\details{
Usually, you will not call this function directly, but will pass it to
a method that works with tour paths like \code{\link[tourr:animate]{tour::animate()}},
\code{\link[tourr:save_history]{tourr::save_history()}} or \code{\link[tourr:render]{tourr::render()}}.
}
\examples{
library(tourr)
twod <- save_history(flea[, 1:3], max = 5)
str(twod)
tourr::animate_xy(flea[, 1:3], woylier::planned_tour_givens(twod))
tourr::animate_xy(flea[, 1:3], woylier::planned_tour_givens(twod, TRUE))
oned <- tourr::save_history(flea[, 1:6], tourr::grand_tour(1), max = 3)
tourr::animate_dist(flea[, 1:6], woylier::planned_tour_givens(oned))
}
\seealso{
The \code{\link[tourr:little_tour]{tourr::little_tour()}}, a special type of planned tour
which cycles between all axis parallel projections.
}
\keyword{dynamic}
\keyword{hplot}
