% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{preprojection}
\alias{preprojection}
\title{Build a d-dimensional pre-projection space by orthonormalizing Fz with regard to Fa}
\usage{
preprojection(Fa, Fz)
}
\arguments{
\item{Fa}{starting pxd frame}

\item{Fz}{ending pxd frame}
}
\value{
B pre-projection px2d matrix
}
\description{
Build a d-dimensional pre-projection space by orthonormalizing Fz with regard to Fa
}
\keyword{internal}
