% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.R
\name{row_rot}
\alias{row_rot}
\title{Takes i and k-th row of a matrix and rotate matrix by theta angle (requires matrix a to be 2*q matrix)}
\usage{
row_rot(a, i, k, theta)
}
\arguments{
\item{a}{matrix}

\item{i}{row}

\item{k}{row that we want to zero the element}

\item{theta}{angle between them}
}
\value{
rotated matrix a
refer to Algorithm 5.1.6 of Matrix computation (Golub, Van)
}
\description{
Takes i and k-th row of a matrix and rotate matrix by theta angle (requires matrix a to be 2*q matrix)
}
\keyword{internal}
