% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woylier-package.R
\docType{package}
\name{woylier-package}
\alias{woylier}
\alias{woylier-package}
\title{woylier: Interpolation between frames}
\description{
Interpolates between frames via Givens path.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
